refactor: 타입 구조 정리 및 중복 제거 (node, spatial, interaction)
