+++ frontend/src/features/mindmap/types/interaction.ts
@@ -1,0 +1,7 @@
+import { NodeId } from "@/features/mindmap/types/node";
+
+export type InteractionMode = "idle" | "potential_drag" | "dragging" | "panning";
+
+export type BaseNodeInfo = {
+    targetId: NodeId | null;
+};
