+++ frontend/src/features/mindmap/utils/MindmapInteractionManager.ts
@@ -1,7 +1,8 @@
 import { MOUSE_DOWN } from "@/constants/mouse";
 import { ATTRIBUTE_NAME_OF_NODE_ID } from "@/features/mindmap/constants/node";
-import { BaseNodeInfo, InteractionMode, ViewportTransform } from "@/features/mindmap/types/mindmap_interaction_type";
-import { NodeId } from "@/features/mindmap/types/mindmapType";
+import { BaseNodeInfo, InteractionMode } from "@/features/mindmap/types/interaction";
+import { NodeId } from "@/features/mindmap/types/node";
+import { ViewportTransform } from "@/features/mindmap/types/spatial";
 import TreeContainer from "@/features/mindmap/utils/TreeContainer";
 import { calcDistance } from "@/utils/calc_distance";
