--- frontend/src/features/quad_tree/types/point.ts
@@ -1,4 +1,0 @@
-export type Point = {
-    x: number;
-    y: number;
-};
