--- frontend/src/features/quad_tree/types/rect.ts
@@ -1,6 +1,0 @@
-export type Rect = {
-    minY: number;
-    maxY: number;
-    minX: number;
-    maxX: number;
-};
