+++ frontend/src/shared/utils/rect_helper.ts
@@ -1,5 +1,4 @@
-import { Point } from "@/features/quad_tree/types/point";
-import { Rect } from "@/features/quad_tree/types/rect";
+import { Point, Rect } from "@/features/mindmap/types/spatial";
 
 /** 특정 점이 지정된 사각형 영역 안에 포함되는지 확인 */
 export const isPointInRect = (point: Point, rect: Rect): boolean => {
